+++ succ/modules/enhanced-conditions/enhanced-conditions.js
@@ -24,20 +24,21 @@ export class EnhancedConditions {
      */
     static async _onReady() {
         if (game.user.isGM) {
+            await EnhancedConditions.loadConditionConfigMap();
             await EnhancedConditions.updateConditionMapFromDefaults();
         }
 
-        let defaultMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultMap);
+        let defaultConditions = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultConditions);
         let conditionMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.map);
 
         const mapType = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.mapType);
         const defaultMapType = Sidekick.getKeyByValue(BUTLER.DEFAULT_CONFIG.enhancedConditions.mapTypes, BUTLER.DEFAULT_CONFIG.enhancedConditions.mapTypes.default);
 
-        // If there's no defaultMap, check storage then set appropriately
-        if (!defaultMap || (Object.keys(defaultMap).length === 0 && defaultMap.constructor === Object)) {
+        // If there's no defaultConditions, check storage then set appropriately
+        if (!defaultConditions || defaultConditions.length === 0) {
             if (game.user.isGM) {
-                defaultMap = await EnhancedConditions._loadDefaultMap();
-                Sidekick.setSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultMap, defaultMap);
+                defaultConditions = await EnhancedConditions._loadDefaultConditions();
+                Sidekick.setSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultConditions, defaultConditions);
             }
         }
 
@@ -48,8 +49,8 @@ export class EnhancedConditions {
 
         // If there's no condition map, get the default one
         if (!conditionMap.length) {
-            // Pass over defaultMap since the storage version is still empty
-            conditionMap = EnhancedConditions.getDefaultMap(defaultMap);
+            // Pass over defaultConditions since the storage version is still empty
+            conditionMap = EnhancedConditions.getMapForDefaultConditions(defaultConditions);
 
             if (game.user.isGM) {
                 const preparedMap = EnhancedConditions._prepareMap(conditionMap);
@@ -81,7 +82,7 @@ export class EnhancedConditions {
             game.succ.conditions = conditionMap;
         }
 
-        Hooks.callAll('succReady', game.succ);  
+        Hooks.callAll('succReady', game.succ);
     }
 
     /**
@@ -91,18 +92,41 @@ export class EnhancedConditions {
         if (game.userId !== userId) {
             return;
         }
-        
-        //Add Conviction:
-        if (update?.system?.details?.conviction?.active === true) {
-            //Check if condition was toggled on token, otherwise toggle it here:
-            if (await EnhancedConditionsAPI.hasCondition('conviction', actor, {warn: false}) === true) { return; }
 
-            await EnhancedConditionsAPI.addCondition('conviction', actor);
+        const convictionCondition = EnhancedConditions.lookupConditionByOption("conviction");
+        if (convictionCondition) {
+            //Add Conviction:
+            if (update?.system?.details?.conviction?.active === true) {
+                //Check if condition was toggled on token, otherwise toggle it here:
+                if (await EnhancedConditionsAPI.hasCondition(convictionCondition.id, actor, { warn: false }) === true) { return; }
+    
+                await EnhancedConditionsAPI.addCondition(convictionCondition.id, actor);
+            }
+    
+            //Remove Conviction:
+            if (update?.system?.details?.conviction?.active === false) {
+                await EnhancedConditionsAPI.removeCondition(convictionCondition.id, actor, { warn: false });
+            }
         }
+    }
 
-        //Remove Conviction:
-        if (update?.system?.details?.conviction?.active === false) {
-            await EnhancedConditionsAPI.removeCondition('conviction', actor, {warn:false});
+    static async _onSwadeActorPrepareDerivedData(actor) {
+        //Get the user defined ID for the condition to be added if overencumbered:
+        const encumberedId = game.succ.conditions.find(c => c.options.encumbered)?.id
+        //Add/Remove Encumbrance if game setting for that is true and a condition is set for it:
+        if (game.settings.get('swade', 'applyEncumbrance') && encumberedId) {
+            //Check if overencumbered and if the condition is already applied:
+            const hasEncumberedCondition = await EnhancedConditionsAPI.hasCondition(encumberedId, actor)
+            let isEncumbered = actor.system.details.encumbrance.value > actor.system.details.encumbrance.max ? true : false
+            //If BRSW is enabled also check if NPCs should get encumbrance:
+            if (game.modules.get('betterrolls-swade2')?.active) {
+                const npcAvoidEncumbrance = game.settings
+                    .get('betterrolls-swade2', 'optional_rules_enabled')
+                    .indexOf('NPCDontUseEncumbrance') > -1;
+                if (npcAvoidEncumbrance && actor.type === 'npc') { isEncumbered = false }
+            }
+            if (isEncumbered && !hasEncumberedCondition) { EnhancedConditionsAPI.addCondition(encumberedId, actor) }
+            else if (!isEncumbered && hasEncumberedCondition) { EnhancedConditionsAPI.removeCondition(encumberedId, actor) }
         }
     }
 
@@ -174,7 +198,7 @@ export class EnhancedConditions {
 
             const actor = ChatMessage.getSpeakerActor(message.speaker);
 
-            EnhancedConditionsAPI.removeCondition(conditionId, actor, {warn: false});
+            EnhancedConditionsAPI.removeCondition(conditionId, actor, { warn: false });
         });
 
         undoRemoveAnchor.on("click", event => {
@@ -221,7 +245,7 @@ export class EnhancedConditions {
 
         effectIcons.each((index, element) => {
             const url = new URL(element.src);
-            const path = url?.pathname?.substring(1); 
+            const path = url?.pathname?.substring(1);
             const conditions = EnhancedConditions.getConditionsByIcon(path);
             const statusEffect = CONFIG.statusEffects.find(e => e.icon === path);
 
@@ -243,9 +267,9 @@ export class EnhancedConditions {
      * @param {ActiveEffect} effect  the effect
      * @param {String} type  the type of change to process
      */
-    static _processActiveEffectChange(effect, type="create", userId) {
+    static _processActiveEffectChange(effect, type = "create", userId) {
         if (!(effect instanceof ActiveEffect)) return;
-        
+
         const effectId = effect.getFlag(`${BUTLER.NAME}`, `${BUTLER.FLAGS.enhancedConditions.conditionId}`);
         if (!effectId) return;
 
@@ -256,8 +280,8 @@ export class EnhancedConditions {
         const shouldOutput = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.outputChat) && condition.options.outputChat;
         const outputType = type === "delete" ? "removed" : "added";
         const actor = effect.parent;
-        
-        if (shouldOutput) EnhancedConditions.outputChatMessage(actor, condition, {type: outputType});
+
+        if (shouldOutput) EnhancedConditions.outputChatMessage(actor, condition, { type: outputType });
         let macros = [];
 
         switch (type) {
@@ -266,7 +290,7 @@ export class EnhancedConditions {
 
                 const hasEffectOptions = hasProperty(effect, `flags.${BUTLER.NAME}.${BUTLER.FLAGS.enhancedConditions.effectOptions}`);
                 if (hasEffectOptions && Object.keys(effect.flags.succ.effectOptions).length > 0) {
-                    EnhancedConditions.applyEffectOptions(effect, actor);                  
+                    EnhancedConditions.applyEffectOptions(effect, actor);
                 } else {
                     EnhancedConditions.applyConditionOptions(condition, actor, "create");
                 }
@@ -286,21 +310,22 @@ export class EnhancedConditions {
         if (macroIds?.length) EnhancedConditions._processMacros(macroIds, actor);
     }
 
-    static applyConditionOptions(condition, actor, type="create") {
+    static applyConditionOptions(condition, actor, type = "create") {
         switch (type) {
             case "create":
                 if (condition.options?.removeOthers) EnhancedConditions.removeOtherConditions(actor, condition.id);
-                if (condition.options?.markDefeated) EnhancedConditions.toggleDefeated(actor, {markDefeated: true});
+                if (condition.options?.markDefeated) EnhancedConditions.toggleDefeated(actor, { markDefeated: true });
                 if (condition.options?.boostTrait) EnhancedConditionsPowers.boostLowerTrait(actor, condition, true);
                 if (condition.options?.lowerTrait) EnhancedConditionsPowers.boostLowerTrait(actor, condition, false);
                 if (condition.options?.smite) EnhancedConditionsPowers.smite(actor, condition);
                 if (condition.options?.protection) EnhancedConditionsPowers.protection(actor, condition);
+                if (condition.options?.deflection) EnhancedConditionsPowers.deflection(actor, condition);
                 if (condition.options?.conviction) EnhancedConditions.activateConviction(actor);
-                
+
                 break;
 
             case "delete":
-                if (condition.options?.markDefeated) EnhancedConditions.toggleDefeated(actor, {markDefeated: false});
+                if (condition.options?.markDefeated) EnhancedConditions.toggleDefeated(actor, { markDefeated: false });
                 if (condition.options?.conviction) EnhancedConditions.deactivateConviction(actor);
                 break;
 
@@ -322,7 +347,7 @@ export class EnhancedConditions {
         //Local function to apply the options that are shared between all effects
         async function applySharedOptions(options) {
             let updates = activeEffect.toObject();
-            foundry.utils.mergeObject(updates.flags, options.flags, {overwrite: false});
+            foundry.utils.mergeObject(updates.flags, options.flags, { overwrite: false });
             if (options.icon) { updates.icon = options.icon; }
             await activeEffect.update(updates);
         }
@@ -350,15 +375,21 @@ export class EnhancedConditions {
             await applySharedOptions(options);
             EnhancedConditionsPowers.protectionBuilder(activeEffect, options.bonus, options.type);
         }
+
+        if (effect.flags.succ.effectOptions.deflection) {
+            let options = effect.flags.succ.effectOptions.deflection;
+            await applySharedOptions(options);
+            EnhancedConditionsPowers.deflectionBuilder(activeEffect, options.type);
+        }
     }
-        
+
 
     /**
      * Checks statusEffect icons against map and returns matching condition mappings
      * @param {Array | String} effectIds  A list of effectIds, or a single effectId to check
      * @param {Array} [map=[]]  the condition map to look in
      */
-    static lookupEntryMapping(effectIds, map=[]) {
+    static lookupEntryMapping(effectIds, map = []) {
         if (!(effectIds instanceof Array)) {
             effectIds = [effectIds];
         }
@@ -371,7 +402,7 @@ export class EnhancedConditions {
         const conditionEntries = map.filter(row => effectIds.includes(row.id ?? Sidekick.generateUniqueSlugId(row.name)));
 
         if (conditionEntries.length === 0) return;
-        
+
         return conditionEntries.length > 1 ? conditionEntries : conditionEntries.shift();
     }
 
@@ -379,7 +410,7 @@ export class EnhancedConditions {
      * Output one or more condition entries to chat
      * @todo refactor to use actor or token
      */
-    static async outputChatMessage(entity, entries, options={type: "active"}) {
+    static async outputChatMessage(entity, entries, options = { type: "active" }) {
         const isActorEntity = entity instanceof Actor;
         const isTokenEntity = entity instanceof Token || entity instanceof TokenDocument;
         // Turn a single condition mapping entry into an array
@@ -410,7 +441,7 @@ export class EnhancedConditions {
         const chatUser = game.userId;
         //const token = token || this.currentToken;
         const chatType = CONST.CHAT_MESSAGE_TYPES.OTHER;
-        const speaker = isActorEntity ? ChatMessage.getSpeaker({actor: entity}) : ChatMessage.getSpeaker({token: entity});
+        const speaker = isActorEntity ? ChatMessage.getSpeaker({ actor: entity }) : ChatMessage.getSpeaker({ token: entity });
         const timestamp = type.active ? null : Date.now();
 
         // iterate over the entries and mark any with references for use in the template
@@ -438,7 +469,7 @@ export class EnhancedConditions {
         const lastMessageSpeaker = lastMessage?.speaker;
         const sameSpeaker = isActorEntity ? lastMessageSpeaker?.actor === speaker.actor : lastMessageSpeaker?.token === speaker.token;
         const hasPermissions = game.user.isGM || lastMessage?.user?.id == game.userId;
-        
+
         // hard code the recent timestamp to 30s for now
         const recentTimestamp = Date.now() <= lastMessage?.timestamp + 30000;
         const enhancedConditionsDiv = lastMessage?.content.match("enhanced-conditions");
@@ -446,14 +477,14 @@ export class EnhancedConditions {
         if (!type.active && enhancedConditionsDiv && sameSpeaker && recentTimestamp && hasPermissions) {
             let newContent = "";
             for (const condition of entries) {
-                const newRow = await renderTemplate(BUTLER.DEFAULT_CONFIG.enhancedConditions.templates.chatConditionsPartial, {condition, type, timestamp});
+                const newRow = await renderTemplate(BUTLER.DEFAULT_CONFIG.enhancedConditions.templates.chatConditionsPartial, { condition, type, timestamp });
                 newContent += newRow;
             }
             const existingContent = lastMessage.content;
             const ulEnd = existingContent?.indexOf(`</ul>`);
             if (!ulEnd) return;
             const content = existingContent.slice(0, ulEnd) + newContent + existingContent.slice(ulEnd);
-            await lastMessage.update({content});
+            await lastMessage.update({ content });
             EnhancedConditions.updateConditionTimestamps();
             ui.chat.scrollBottom();
         } else {
@@ -473,7 +504,7 @@ export class EnhancedConditions {
      * @param {Actor | Token} entities  the entity to mark defeated
      * @param {Boolean} options.markDefeated  an optional state flag (default=true)
      */
-    static toggleDefeated(entities, {markDefeated=true}={}) {
+    static toggleDefeated(entities, { markDefeated = true } = {}) {
         const combat = game.combat;
 
         if (!entities) {
@@ -481,7 +512,7 @@ export class EnhancedConditions {
             if (canvas?.tokens?.controlled.length) entities = canvas.tokens.controlled;
             else if (game.user.character) entities = game.user.character;
         }
-        
+
         if (!entities) {
             return;
         }
@@ -494,7 +525,7 @@ export class EnhancedConditions {
 
         // loop through tokens, and if there's matching combatants, add them to the update
         for (const token of tokens) {
-            
+
             const combatants = combat ? combat.combatants?.contents?.filter(c => c.tokenId === token.id && c.defeated != markDefeated) : [];
 
             if (!combatants.length) return;
@@ -527,7 +558,7 @@ export class EnhancedConditions {
         } else if (actor.system.details.conviction.value < 1 && actor.system.details.conviction.active === false) {
             //Condition was toggled instead of the button on the actor sheet but actor has no conviction tokens.
             ui.notifications.warn(game.i18n.localize("ENHANCED_CONDITIONS.Notification.NoConvictionToken"));
-            await EnhancedConditionsAPI.removeCondition('conviction', actor, {warn: true});
+            await EnhancedConditionsAPI.removeCondition('conviction', actor, { warn: true });
         }
     }
 
@@ -549,7 +580,7 @@ export class EnhancedConditions {
      * @param {*} conditionId 
      */
     static async removeOtherConditions(entity, conditionId) {
-        const entityConditions = await EnhancedConditionsAPI.getConditions(entity, {warn: false});
+        const entityConditions = await EnhancedConditionsAPI.getConditions(entity, { warn: false });
         let conditions = entityConditions ? entityConditions.conditions : [];
         conditions = conditions instanceof Array ? conditions : [conditions];
 
@@ -559,7 +590,7 @@ export class EnhancedConditions {
 
         if (!removeConditions.length) return;
 
-        for (const c of removeConditions) await EnhancedConditionsAPI.removeCondition(c.id, entity, {warn: true});
+        for (const c of removeConditions) await EnhancedConditionsAPI.removeCondition(c.id, entity, { warn: true });
     }
 
     /**
@@ -567,7 +598,7 @@ export class EnhancedConditions {
      * @param {*} macroIds 
      * @param {*} entity 
      */
-    static async _processMacros(macroIds, entity=null) {
+    static async _processMacros(macroIds, entity = null) {
         const isToken = entity instanceof Token || entity instanceof TokenDocument;
         const isActor = entity instanceof Actor;
 
@@ -575,7 +606,7 @@ export class EnhancedConditions {
             const macro = game.macros.get(macroId);
             if (!macro) continue;
 
-            const scope = isToken ? {token: entity} : (isActor ? {actor: entity} : null);
+            const scope = isToken ? { token: entity } : (isActor ? { actor: entity } : null);
             await macro.execute(scope);
         }
     }
@@ -585,7 +616,7 @@ export class EnhancedConditions {
      */
     static updateConditionTimestamps() {
         const conditionRows = document.querySelectorAll("ol#chat-log ul.condition-list li");
-        for ( const li of conditionRows ) {
+        for (const li of conditionRows) {
             const timestamp = typeof li.dataset.timestamp === "string" ? parseInt(li.dataset.timestamp) : li.dataset.timestamp;
             const iconSpanWrapper = li.querySelector("span.add-remove-icon");
 
@@ -614,7 +645,7 @@ export class EnhancedConditions {
                     <i class="fas fa-flask"></i> ${BUTLER.DEFAULT_CONFIG.enhancedConditions.conditionLab.title}
                 </button>`
         );
-        
+
         succDiv.append(labButton);
 
         labButton.on("click", event => {
@@ -636,74 +667,48 @@ export class EnhancedConditions {
 
         if (display && labButton && labButton.style.display !== "block") {
             return labButton.style.display = "block";
-        } 
-        
+        }
+
         if (labButton && !display && labButton.style.display !== "none") {
             return labButton.style.display = "none";
         }
     }
 
     /**
-     * Returns the default maps supplied with the module
-     * 
-     * @todo: map to entryId and then rebuild on import
+     * Loads teh condition map json and applies any system overrides
      */
-    static async _loadDefaultMap() {
+    static async loadConditionConfigMap() {
         const source = "data";
-        const conditionMapFilePath = BUTLER.DEFAULT_CONFIG.enhancedConditions.conditionMapFilePath;
-        const overridesPath = BUTLER.DEFAULT_CONFIG.enhancedConditions.conditionModuleOverridesPath;
-        const conditionMapJson = await Sidekick.fetchJson(conditionMapFilePath);
-        const overridesJsons = await Sidekick.fetchJsons(source, overridesPath);
-        
-        const defaultMap = conditionMapJson.map;
-        
-        //Loop over our overrides and check if any of them are active
-        for (let overrides of overridesJsons) {
-            if (game.modules.get(overrides.module)?.active) {
-                for (const override of overrides.map) {
-                    const condition = defaultMap.find(c => c.id === override.id);
-                    if (!condition) {
-                        //If the condition doesn't exist, add it to the map
-                        defaultMap.push(override);
-                    } else {
-                        //If the condition exists, merge the data instead
-                        foundry.utils.mergeObject(condition, override);
-                    }
-                }
-                break;
-            }
-        }
-        
-        // If the default config contains changes and we have not overridden them in the system definition, copy those over
-        const statusEffects = CONFIG.defaultStatusEffects ? CONFIG.defaultStatusEffects : CONFIG.statusEffects;
-        for (let statusEffect of statusEffects) {
-            if (!statusEffect.changes && !statusEffect.duration && !statusEffect.flags) {
-                continue;
-            }
+        const overridesJsons = await Sidekick.fetchJsons(source, BUTLER.DEFAULT_CONFIG.enhancedConditions.conditionModuleOverridesPath);
+        const conditionConfigJson = await Sidekick.fetchJson(BUTLER.DEFAULT_CONFIG.enhancedConditions.conditionConfigFilePath);
+        const groupsJsons = await Sidekick.fetchJsons(source, BUTLER.DEFAULT_CONFIG.enhancedConditions.defaultConditionGroupsPath);
+        game.succ.conditionConfigMap = conditionConfigJson.map;
 
-            let defaultCondition = defaultMap.find(c => c.id === statusEffect.id);
-            if (defaultCondition) {
-                if (defaultCondition.ignoreSystemSettings) {
-                    // We've decided that we want to prioritize what we have in our condition-map.json file
+        let defaultConditions = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultConditions);
+        if (!defaultConditions || defaultConditions.length === 0) {
+            defaultConditions = undefined;
+        }
+
+        for (let condition of game.succ.conditionConfigMap) {
+            for (let group of groupsJsons) {
+                if (group.canBeDisabled) {
                     continue;
                 }
 
-                defaultCondition.activeEffect = defaultCondition.activeEffect ? defaultCondition.activeEffect : {};
-
-                // Delete any existing data before using what's in the system
-                // This is to avoid cases where, for example, our config had a duration but the system's does not
-                // We want to ensure that, if the system has changes, we are matching it exactly 
-                delete defaultCondition.activeEffect.changes;
-                delete defaultCondition.activeEffect.duration;
-                delete defaultCondition.activeEffect.flags;
-
-                defaultCondition.activeEffect = {
-                    ...defaultCondition.activeEffect,
-                    ...statusEffect.changes != undefined ? {changes: statusEffect.changes} : null,
-                    ...statusEffect.duration != undefined ? {duration: statusEffect.duration} : null,
-                    ...statusEffect.flags != undefined ? {flags: statusEffect.flags} : null
+                let conditionId = group.conditions.find(c => c === condition.id);
+                if (conditionId) {
+                    condition.destroyDisabled = true;
+                    break;
                 }
-            } else {
+            }
+        }
+
+        //Loop over the default conditions and look for ones that are missing from our full map
+        const statusEffects = CONFIG.defaultStatusEffects ? CONFIG.defaultStatusEffects : CONFIG.statusEffects;
+        for (let statusEffect of statusEffects) {
+            const conditionConfig = game.succ.conditionConfigMap.find(c => c.id === statusEffect.id);
+            if (!conditionConfig) {
+                //If the condition doesn't exist in the full map, it must be something new that was added to the system, so we need to add it
                 let newCondition = {
                     id: statusEffect.id,
                     name: statusEffect.label,
@@ -717,12 +722,59 @@ export class EnhancedConditions {
                         flags: statusEffect.flags
                     };
                 }
-
-                defaultMap.push(newCondition);
+                game.succ.conditionConfigMap.push(newCondition);
             }
         }
 
-        for (let condition of defaultMap) {
+        //Loop over our overrides and check if any of them are active
+        overridesJsons.sort((a, b) => a.priority - b.priority);
+        for (let overrides of overridesJsons) {
+            if (game.modules.get(overrides.module)?.active) {
+                for (const override of overrides.map) {
+                    const condition = game.succ.conditionConfigMap.find(c => c.id === override.id);
+                    if (!condition) {
+                        //If the condition doesn't exist, add it to the map
+                        game.succ.conditionConfigMap.push(override);
+                    } else {
+                        //If the condition exists, merge the data instead
+                        foundry.utils.mergeObject(condition, override);
+                    }
+                }
+            }
+        }
+
+        // If the default config contains changes and we have not overridden them in the system definition, copy those over
+        for (let statusEffect of statusEffects) {
+            if (!statusEffect.changes && !statusEffect.duration && !statusEffect.flags) {
+                continue;
+            }
+
+            let conditionConfig = game.succ.conditionConfigMap.find(c => c.id === statusEffect.id);
+            if (conditionConfig) {
+                if (conditionConfig.ignoreSystemSettings) {
+                    // We've decided that we want to prioritize what we have in our condition-config.json file
+                    continue;
+                }
+
+                conditionConfig.activeEffect = conditionConfig.activeEffect ? conditionConfig.activeEffect : {};
+
+                // Delete any existing data before using what's in the system
+                // This is to avoid cases where, for example, our config had a duration but the system's does not
+                // We want to ensure that, if the system has changes, we are matching it exactly 
+                delete conditionConfig.activeEffect.changes;
+                delete conditionConfig.activeEffect.duration;
+                delete conditionConfig.activeEffect.flags;
+
+                conditionConfig.activeEffect = {
+                    ...conditionConfig.activeEffect,
+                    ...statusEffect.changes != undefined ? { changes: statusEffect.changes } : null,
+                    ...statusEffect.duration != undefined ? { duration: statusEffect.duration } : null,
+                    ...statusEffect.flags != undefined ? { flags: statusEffect.flags } : null
+                }
+            }
+        }
+
+        for (let condition of game.succ.conditionConfigMap) {
             if (condition.referenceId) {
                 let regex = /(?<=\{).+(?=\})/;
                 let match = condition.referenceId.match(regex);
@@ -734,13 +786,43 @@ export class EnhancedConditions {
                     condition.referenceId += `{${game.i18n.localize(condition.name)}}`;
                 }
             }
-            
+
             if (condition.activeEffect) {
                 condition.activeEffect.label = game.i18n.localize(condition.name);
             }
         }
+    }
 
-        return defaultMap;
+    /**
+     * Returns the default maps supplied with the module
+     */
+    static async _loadDefaultConditions() {
+        const source = "data";
+        const groupsJsons = await Sidekick.fetchJsons(source, BUTLER.DEFAULT_CONFIG.enhancedConditions.defaultConditionGroupsPath);
+
+        const defaultConditions = [];
+
+        for (let condition of game.succ.conditionConfigMap) {
+            let foundCondition = false;
+            for (let group of groupsJsons) {
+                let conditionId = group.conditions.find(c => c === condition.id);
+                if (conditionId) {
+                    foundCondition = true;
+                    if (group.enabledByDefault) {
+                        defaultConditions.push(condition.id);
+                    }
+                    break;
+                }
+            }
+
+            if (!foundCondition) {
+                //This condition doesn't exist in any of our groups so we'll assume it came from the system
+                //Add it to the default map so that the user can see it
+                defaultConditions.push(condition.id);
+            }
+        }
+
+        return defaultConditions;
     }
 
     /**
@@ -759,7 +841,7 @@ export class EnhancedConditions {
         // Map existing ids for ease of access
         const existingIds = conditionMap.filter(c => c.id).map(c => c.id);
         const processedIds = [];
-        
+
         // Iterate through the map validating/preparing the data
         for (let i = 0; i < conditionMap.length; i++) {
             let condition = duplicate(conditionMap[i]);
@@ -805,7 +887,7 @@ export class EnhancedConditions {
     /**
      * Duplicate the core special status effect mappings, freeze the duplicate then store a copy in settings
      */
-     static _backupCoreSpecialStatusEffects() {
+    static _backupCoreSpecialStatusEffects() {
         CONFIG.defaultSpecialStatusEffects = CONFIG.defaultSpecialStatusEffects || foundry.utils.duplicate(CONFIG.specialStatusEffects);
         if (!Object.isFrozen(CONFIG.defaultSpecialStatusEffects)) {
             Object.freeze(CONFIG.defaultSpecialStatusEffects);
@@ -818,7 +900,7 @@ export class EnhancedConditions {
      * @param {String} conditionId  the condition to get
      * @param {Array} map  the condition map to search
      */
-    static lookupConditionById(conditionId, map=null) {
+    static lookupConditionById(conditionId, map = null) {
         if (!conditionId) return;
 
         conditionId = conditionId instanceof Array ? conditionId : [conditionId];
@@ -830,6 +912,18 @@ export class EnhancedConditions {
         if (!conditions.length) return null;
 
         return conditions.length > 1 ? conditions : conditions.shift();
+    }
+
+    /**
+     * Gets one or more conditions from the map by their name
+     * @param {String} conditionId  the condition to get
+     * @param {Array} map  the condition map to search
+     */
+    static lookupConditionByOption(option) {
+        if (!option) return;
+
+        const map = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.map);
+        return map.find(c => c.options && c.options[option]);
     }
 
     /**
@@ -873,11 +967,11 @@ export class EnhancedConditions {
                 return;
             }
 
-            const defaultCondition = activeConditionMap.find(c => c.id === CONFIG.specialStatusEffects[key]);
-            if (defaultCondition) {
+            const activeCondition = activeConditionMap.find(c => c.id === CONFIG.specialStatusEffects[key]);
+            if (activeCondition) {
                 //The name of this effect matches one of our effect. Check its options
-                if (hasProperty(defaultCondition.options, optionName)) {
-                    if (getProperty(defaultCondition.options, optionName)) {
+                if (hasProperty(activeCondition.options, optionName)) {
+                    if (getProperty(activeCondition.options, optionName)) {
                         //This effect has the option enabled, so just leave it as is
                         return;
                     }
@@ -914,9 +1008,9 @@ export class EnhancedConditions {
         if (!conditionMap.length) return;
 
         const existingIds = conditionMap.filter(c => c.id).map(c => c.id);
-        
+
         const statusEffects = [];
-        
+
         for (const c of conditionMap) {
             const id = c.id ?? Sidekick.createId(existingIds);
             const effect = {
@@ -956,19 +1050,19 @@ export class EnhancedConditions {
                 effect.flags.core.overlay = overlay;
             }
         }
-        
+
         return effects;
     }
 
     /**
      * Returns just the icon side of the map
      */
-    static getConditionIcons(conditionMap={}) {
+    static getConditionIcons(conditionMap = {}) {
         if (!conditionMap) {
             //maybe log an error?
             return;
         }
-        
+
         if (Object.keys(conditionMap).length === 0) {
             conditionMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.map);
 
@@ -976,7 +1070,7 @@ export class EnhancedConditions {
                 return [];
             }
         }
-        
+
         if (conditionMap instanceof Array) {
             return conditionMap.map(mapEntry => mapEntry.icon);
         }
@@ -988,7 +1082,7 @@ export class EnhancedConditions {
      * Retrieves a condition icon by its mapped name
      * @param {*} condition 
      */
-    static getIconsByCondition(condition, {firstOnly=false}={}) {
+    static getIconsByCondition(condition, { firstOnly = false } = {}) {
         const conditionMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.map);
 
         if (!conditionMap || !condition) {
@@ -1011,7 +1105,7 @@ export class EnhancedConditions {
      * Retrieves a condition name by its mapped icon
      * @param {*} icon 
      */
-    static getConditionsByIcon(icon, {firstOnly=false}={}) {
+    static getConditionsByIcon(icon, { firstOnly = false } = {}) {
         const conditionMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.map);
 
         if (!conditionMap || !icon) {
@@ -1037,90 +1131,81 @@ export class EnhancedConditions {
         if (json.system !== game.system.id) {
             ui.notifications.warn(game.i18n.localize("ENHANCED_CONDITIONS.MapMismatch"));
         }
-        
+
         const map = json.map ? EnhancedConditions._prepareMap(json.map) : [];
 
         return map;
     }
 
     /**
-     * Returns the default condition map for a given system
+     * Builds a map of default conditions from a list of conditions
      */
-    static getDefaultMap(defaultMap=null) {
-        defaultMap = defaultMap ? defaultMap : Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultMap);
-
-        if (!defaultMap) {
-            defaultMap = EnhancedConditions.buildDefaultMap();
+    static getMapForDefaultConditions(defaultConditions) {
+        let map = [];
+        for (let conditionId of defaultConditions) {
+            map.push(game.succ.conditionConfigMap.find(c => c.id === conditionId));
         }
-
-        return defaultMap;
+        return map;
     }
 
     /**
-     * Builds a default map for a given system
-     * @todo #281 update for active effects
+     * Returns the default condition map for a given system
      */
-    static buildDefaultMap() {
-        const coreEffectsSetting = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.coreEffects) 
-        const coreEffects = (coreEffectsSetting && coreEffectsSetting.length) ? coreEffectsSetting : CONFIG.statusEffects;
-        const map = EnhancedConditions._prepareMap(coreEffects);
+    static getDefaultMap(defaultConditions = null) {
+        defaultConditions = defaultConditions ? defaultConditions : Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultConditions);
 
-        return map;
+        return EnhancedConditions.getMapForDefaultConditions(defaultConditions);
     }
-    
+
     /**
      * Updates the condition map to include any changes from the default map and system settings
      * If the user has made their own changes to a condition, the condition in the default map will be ignored
      */
     static async updateConditionMapFromDefaults() {
         let conditionMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.map);
-        let defaultMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultMap);
+        let defaultConditions = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultConditions);
         let deletedConditionsMap = Sidekick.getSetting(BUTLER.SETTING_KEYS.enhancedConditions.deletedConditionsMap);
-        if (!conditionMap.length || !defaultMap) {
-            return;
-        }
-        
+
         let currentDiffs = [];
-        for (let defaultCondition of defaultMap) {
-            const condition = conditionMap.find(c => c.name === defaultCondition.name);
+        for (let conditionConfig of game.succ.conditionConfigMap) {
+            const condition = conditionMap.find(c => c.id === conditionConfig.id);
             if (condition) {
-                if (!!condition.activeEffect != !!defaultCondition.activeEffect) {
+                if (!!condition.activeEffect != !!conditionConfig.activeEffect) {
                     //One has an active effect and the other doesn't which means they must be different
-                    currentDiffs.push(defaultCondition);
+                    currentDiffs.push(conditionConfig);
                     continue;
                 }
 
-                if (!defaultCondition.activeEffect) {
+                if (!conditionConfig.activeEffect) {
                     //Both must be null so that means they are the same
                     continue;
                 }
 
-                if (JSON.stringify(condition.activeEffect.changes) !== JSON.stringify(defaultCondition.activeEffect.changes) ||
-                    JSON.stringify(condition.activeEffect.flags) !== JSON.stringify(defaultCondition.activeEffect.flags) ||
-                    JSON.stringify(condition.activeEffect.duration) !== JSON.stringify(defaultCondition.activeEffect.duration)) {
-                    currentDiffs.push(defaultCondition);
+                if (JSON.stringify(condition.activeEffect.changes) !== JSON.stringify(conditionConfig.activeEffect.changes) ||
+                    JSON.stringify(condition.activeEffect.flags) !== JSON.stringify(conditionConfig.activeEffect.flags) ||
+                    JSON.stringify(condition.activeEffect.duration) !== JSON.stringify(conditionConfig.activeEffect.duration)) {
+                    currentDiffs.push(conditionConfig);
                 }
             }
         }
 
-        defaultMap = await EnhancedConditions._loadDefaultMap();
-
-        for (let defaultCondition of defaultMap) {
-            const condition = conditionMap.find(c => c.name === defaultCondition.name);
+        for (let conditionConfig of game.succ.conditionConfigMap) {
+            const condition = conditionMap.find(c => c.id === conditionConfig.id);
             if (!condition) {
-                const deletedCondition = deletedConditionsMap.find(c => c.name === defaultCondition.name);
-                if (!deletedCondition) {
-                    conditionMap.push(duplicate(defaultCondition));
+                const defaultCondition = defaultConditions.find(c => c === conditionConfig.id);
+                const deletedCondition = deletedConditionsMap.find(c => c.id === conditionConfig.id);
+                if (defaultCondition && !deletedCondition) {
+                    conditionMap.push(duplicate(conditionConfig));
                 }
             } else {
-                if (!defaultCondition.activeEffect) {
+                if (!conditionConfig.activeEffect) {
                     //The default condition doesn't have an active effect, so we'll just leave whatever is in the condition alone
                     //We're making the assumption here that none of the core system effects will ever have an active effect in one version and then remove it later
                     //It's very unlikely to ever happen and in the rare case it does, the user can manually reset their map to fix it
                     continue;
                 }
 
-                let existingDiff = currentDiffs.find(d => d.name === defaultCondition.name);
+                let existingDiff = currentDiffs.find(d => d.id === conditionConfig.id);
                 if (existingDiff) {
                     //If we had an existing diff for this condition, we'll assume the user wants it that way and leave it as is 
                     continue;
@@ -1128,11 +1213,10 @@ export class EnhancedConditions {
 
                 //If we didn't have an existing diff, we'll assume the user wants the most up to date functionality
                 //For simplicity, we set the value here regardless of if it's changed or not
-                condition.activeEffect = {...defaultCondition.activeEffect};
+                condition.activeEffect = { ...conditionConfig.activeEffect };
             }
         }
 
-        await Sidekick.setSetting(BUTLER.SETTING_KEYS.enhancedConditions.defaultMap, defaultMap, true);
         await Sidekick.setSetting(BUTLER.SETTING_KEYS.enhancedConditions.map, conditionMap, true);
     }
 }